+++ src/ui/pages/settings.rs
@@ -38,8 +38,7 @@ impl Component for SettingsPage {
         let _ = self.settings; // to appease clippy
         let items = ["1", "2", "3"].map(ListItem::new);
         let list = List::new(items).highlight_style(THEME.selected_text);
-        let mut state =
-            ListState::default().with_selected(Option::from(self.selector.index as usize));
+        let mut state = ListState::default().with_selected(Option::from(self.selector.index));
         frame.render_stateful_widget(list, area, &mut state);
         Ok(())
     }
